refactor: extract endpoint validation helpers
Move validate_discord_webhook_url from alerts.rs into a new
endpoint_validation module and add validate_discord_channel_id as
a sibling helper. The webhook URL tests move along with the function
so they continue to exercise the moved code path.

This is a pure refactor in preparation for the endpoints API
introduced in Task 4 of the notification tier 1 plan.
