+++ ultros/src/web/api/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod alerts;
 mod best_deals;
 mod cheapest_per_world;
+pub(crate) mod endpoint_validation;
 pub(crate) mod real_time_data;
 mod recent_sales;
 mod trends;
